Clean some docs and move skfem.models.helpers
