+++ skfem/element/element_hdiv.py
@@ -23,4 +23,4 @@ class ElementHdiv(Element):
         ),)
 
     def lbasis(self, X, i):
-        raise Exception("ElementHdiv lbasis method not found.")
+        raise Exception("ElementHdiv.lbasis method not found.")
